AIFunction/requirements.txt
functions-framework==3.*
numpy==1.24.3
openai==0.27.8
pandas==2.0.2
python-dotenv==0.21.1
matplotlib==3.7.1
plotly==5.15.0
scipy==1.10.1
scikit-learn==1.2.2